tres support in pyot: new features supported; push input resource and multiple output supported

+++ pyot/models/tres.py
@@ -174,20 +174,22 @@ class TResT(models.Model):
     by the emu field.
     """
     pf = models.ForeignKey(TResProcessing, related_name='ProcessingFunction')
-    inputS = models.ManyToManyField(Resource)
-    output = models.ForeignKey(Resource, related_name='OutputDestination',
-                               null=True)
+    inputS = models.ManyToManyField(Resource, blank=True, related_name='is')
+    output = models.ManyToManyField(Resource, blank=True, related_name='od')
     TResResource = models.ForeignKey(Resource, null=True,
                                      related_name='TresResource')
     state = models.CharField(max_length=10, blank=False, choices=TRES_STATES,
                              default='CREATED')
     period = models.IntegerField(default=0)
     emu = models.ForeignKey(EmulatorState, null=True)
+
     class Meta(object):
         app_label = 'pyot'
 
     def __unicode__(self):
-        return u"Pf={p}, inputs={i}, output={o}".format(p=self.pf, i=str(self.inputS.all()), o=self.output)
+        return u"Pf={p}, inputs={i}, output={o}".format(p=self.pf,
+                                                        i=str(self.inputS.all()),
+                                                        o=str(self.output.all()))
 
     def deploy(self, t_res_resource):
         """
@@ -202,37 +204,46 @@ class TResT(models.Model):
 
         # 1) compile the script
         basename = os.path.basename(str(self.pf.sourcefile))
-        compile_command = tresCompile + ' ' + tresPMfeat + ' ' + str(self.pf.sourcefile)
+        compile_command = tresCompile + ' ' + (tresPMfeat + ' '
+                                               + str(self.pf.sourcefile))
         p = subprocess.check_call([compile_command],
-                             stdin=subprocess.PIPE,
-                             stdout=subprocess.PIPE,
-                             stderr=subprocess.PIPE,
-                             shell=True, cwd=SCRIPT_FOLDER)
-
-        # 2) start a task downloading the script from the server
+                                  stdin=subprocess.PIPE,
+                                  stdout=subprocess.PIPE,
+                                  stderr=subprocess.PIPE,
+                                  shell=True, cwd=SCRIPT_FOLDER)
+        # start a task downloading the script from the server
         pycFilename = basename + 'c'
         r = tresDownloadScript.apply_async(args=[pycFilename],
-                                       queue=t_res_resource.host.getQueue())
+                                           queue=t_res_resource.host.getQueue())
         r.wait()
         result = r.result
         if result.code != SUCCESS:
             return Response(FAILURE, 'PWN: Error downloading pyc.')
         try:
-            newTask = Resource.objects.get(host=self.TResResource.host, uri='/tasks/' + self.pf.name)
+            newTask = Resource.objects.get(host=self.TResResource.host,
+                                           uri='/tasks/' + self.pf.name)
         except Resource.DoesNotExist:
-            newTask = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name)
+            newTask = Resource.objects.create(host=self.TResResource.host,
+                                              uri='/tasks/' + self.pf.name)
 
         # 3) Create a new task resource
         r = newTask.PUT(query="per=" + str(self.period))
         print 'first put result = ' + r.code
         if r.code != CREATED:
             newTask.delete()
-            return Response(FAILURE, 'Error creating new resource: ' + '/tasks/' + self.pf.name)
+            return Response(FAILURE, 'Error creating new resource: ' +
+                            '/tasks/' + self.pf.name)
 
-        newIs = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name + '/is')
-        newOd = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name + '/od')
-        newPf = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name + '/pf')
-        _newLo = Resource.objects.create(host=self.TResResource.host, uri='/tasks/' + self.pf.name + '/lo')
+        newIs = Resource.objects.create(host=self.TResResource.host,
+                                        uri='/tasks/' + self.pf.name + '/is')
+        newOd = Resource.objects.create(host=self.TResResource.host,
+                                        uri='/tasks/' + self.pf.name + '/od')
+        newPf = Resource.objects.create(host=self.TResResource.host,
+                                        uri='/tasks/' + self.pf.name + '/pf')
+        _newLo = Resource.objects.create(host=self.TResResource.host,
+                                         uri='/tasks/' + self.pf.name + '/lo')
+        _newIn = Resource.objects.create(host=self.TResResource.host,
+                                         uri='/tasks/' + self.pf.name + '/in')
 
         # 4) Upload the processing function
         r = newPf.PUT(inputfile=tmpDir + '/' + pycFilename, block=64)
@@ -242,10 +253,10 @@ class TResT(models.Model):
             self.uninstall()
             return Response(FAILURE, 'Error uploading processing function.')
 
-        # 5) Create output destination resource
-        if self.output:
-            r = newOd.PUT(payload='<' + self.output.getFullURI() + '>')
-            print 'OD put result = ' + r.code
+        # 5) Create output destination resources
+        for od in self.output.all():
+            r = newOd.POST(payload='<' + od.getFullURI() + '>')
+            print 'OD put result = ' + r.code + ' ' + od.getFullURI()
             if r.code != CHANGED:
                 self.uninstall()
                 return Response(FAILURE, 'Error updating OD resource')
@@ -267,17 +278,27 @@ class TResT(models.Model):
         """
         Uninstalls a T-Res task from a node.
         """
-        newTask = Resource.objects.get(host=self.TResResource.host, uri='/tasks/' + self.pf.name)
+        if self.state == 'RUNNING':
+            self.stop()
+
+        if self.state == 'CLEARED':
+            return Response(SUCCESS, 'Task ' + self.pf.name +
+                            'already uninstalled')
+
+        newTask = Resource.objects.get(host=self.TResResource.host,
+                                       uri='/tasks/' + self.pf.name)
         r = newTask.DELETE()
         if r.code == DELETED:
             newTask.delete()
             Resource.objects.filter(host=self.TResResource.host,
-                                    uri__startswith='/tasks/' + self.pf.name).delete()
+                                    uri__startswith='/tasks/' +
+                                    self.pf.name).delete()
             self.state = 'CLEARED'
             self.save()
             return Response(SUCCESS, 'Task ' + self.pf.name + ' uninstalled')
         else:
-            return Response(FAILURE, 'Error uninstalling task ' + self.pf.name)
+            return Response(FAILURE, 'Error uninstalling task ' +
+                            self.pf.name)
 
     def start(self):
         '''
@@ -315,15 +336,25 @@ class TResT(models.Model):
         """
         raise NotImplementedError("Still to be implemented in T-Res")
 
-
     def getLastOutput(self):
         """
         Returns the last output resource reference. The resource can be
         retrieved or observed.
         """
         last_output = Resource.objects.get(host=self.TResResource.host,
-                                           uri='/tasks/' + self.pf.name + '/lo')
+                                           uri='/tasks/' +
+                                           self.pf.name + '/lo')
         return last_output
+
+    def getInputResource(self):
+        """
+        Returns the input resource reference. The resource can be
+        retrieved or observed.
+        """
+        input_resource = Resource.objects.get(host=self.TResResource.host,
+                                              uri='/tasks/' +
+                                              self.pf.name + '/in')
+        return input_resource
 
     def getInputSource(self):
         """
